Rename `public/` folder to `public/docs` (#2930)
This is prep for Learning migrating to this repository and Next.js app.
For that to happen, our links are going to need to start having a
`/docs` prefix, like `/docs/images` rather than `/images`.

This PR renames the `public/` folder to `public/docs`, and adjusts our
tooling to handle the change. We do _not_ yet start using the new links
of `/docs/images` rather than `/images`, but that will be done in a
follow up. This PR is only meant to split up the change and allow
content writers to start organizing files in the appropriate place.

The closed source app will copy the `public/docs` folder to live at
`public/`, like how this PR updates our local & PR previews to strip the
`/docs` prefix.