Automate auditing new pills (#2829)
Closes #2728 

A new script `checkNewPills.ts` checks the toc in `/guides`,
`/migration-guides`, and `/open-source` for pages with the `isNew :
true` entry and returns an error if any of these entries are over two
weeks old.

The proposed way to track the date in which the page was added is for
the content writer to add an `isNewDate: YYYY-MM-DD` entry to track the
date which the page was (approximately) published. If this date is older
than two weeks (or the `isNewDate` entry is missing), then this check
errors with a list of affected urls in which to remove the new pill
from.

This PR also adds the `checkNewPills.ts` to our weekly cron job in order
to notify us when a new pill becomes outdated.

---------

Co-authored-by: Eric Arellano <14852634+Eric-Arellano@users.noreply.github.com>