Correct preset pass manager options (#3016)
Closes #2894 

I only found a couple places which refer to `StochasticSwap` in our
docs. This pass was superseded by `SabreSwap` in Qiskit v2.0, otherwise
the page lists all options for the `generate_preset_pass_manager()`
function correctly.

I've also opened an issue to fix one of the API docstrings which still
refers to `StochasticSwap` as a valid option:
https://github.com/Qiskit/qiskit/issues/14222