Add script to extract notebook image outputs (#3095)
This PR adds a script to extract image outputs from notebooks and
convert them to AVIF if necessary. We're already doing this in
inner-source as part of the sync but it was taking too long with the new
tutorial notebooks. This PR just introduces the script, a follow-up PR
will extract the images and add this to `./fix`.

---------

Co-authored-by: Rebecca Dimock <beckyd@us.ibm.com>
Co-authored-by: Eric Arellano <14852634+Eric-Arellano@users.noreply.github.com>