Use lock to avoid kernel address conflicts in nb-tester (#3143)
This PR uses a lock to try and avoid this error when running in a GitHub
action:

```
zmq.error.ZMQError: Address already in use (addr='tcp://127.0.0.1:45769')
```

This is a bit of a shot in the dark because I can't reproduce it
locally.

This is related to #2387, but is not necessarily a full fix as I believe
there is more than one thing going on there.

***

Inspired by https://github.com/jupyter/nbclient/pull/327