Fix small error (#3115)
The Hello World guide seems to have a error (or potentially misleading
statement) in its description of an expectation value. While it is true
that the referenced expectation value is 1 for the specific entangled
state created by the example circuit, that expectation value does not
have to be 1 for all entangled states.