[nb-tester]: Handle our cloud instance (#3273)
This adds a feature to map between open backend names and the names used
in our cloud testing instance. This is a temporary workaround until we
get access to a testing instance that uses the open names.

This PR does not enable the feature – we'll need to do that with the
switch to cloud, but I've tested it on a few notebooks using my cloud
account and it seems to work well.