Add check to detect overly long inline katex expressions (#3408)
Inline katex expressions are problematic because they will soon start to
overflow the prose of the docs. This is because we won't be controlling
the overflow on any paragraph anymore to allow the use of tooltips on
them.

For this, we decided to remove the inline katex (`$`) references and
instead use `katex-display` (`$$`) when the expressions do not fit in
our smallest screen supported. Given that this is not easy to foresee,
this PR adds a new check that will verify that all the inline katex
expressions introduced in a file are always going to render without
overflowing.

When a bad expressions is detected, we will see the conflicting
expression and file with a message indicating what we can do to fix it:

- Run `./fix` for the output of the notebooks, or
- Manually replace the inline math (`$`) by equations (`$$`)

In addition, this PR updates all the bad expressions found in our
non-API docs. The current versions of the API docs only had one issue
and it's being addressed in
https://github.com/Qiskit/qiskit-addon-cutting/pull/759

---------

Co-authored-by: Frank Harkins <frankharkins@hotmail.co.uk>
Co-authored-by: Eric Arellano <14852634+Eric-Arellano@users.noreply.github.com>