Stop using the legacy app in previews (#3468)
This PR deletes the `platform` option from the `start` script. This is
because we are removing the legacy app in inner source and from the
preview app. It also simplifies the render checker to only verify cloud
pages.

This change will be accompanied by a follow-up that finishes removing
missing parts once we don't use any platform specific tag in the
content.