Migrate the link checker to use the remark parser (#3593)
This PR migrates the link checker to use the `remark` parser instead of
the `rehype` parser. Rehype is meant to take html as input whereas
Remark expects markdown.

Turns out that the only visitor that was extracting links before was the
one looking for `text` nodes. After this change, we can remove the
visitor and rely on the node types `link` and `mdxJsxTextElement` for
`<a>` tags. It also allows us to remove the `markdown-link-extractor`
dependency.

This new approach also finds links that we weren't checking before (it
even found some broken ones). The previous approach was checking 1638
links on our non-api docs, and after this PR we will check 1651 links.