Fix Qiskit's supported OpenQASM 3 feature table
The notes at the top of the table say that "full support" involves being
able to load from OpenQASM 3, but a couple of the rows indicated "full
support" and used the note to explain that this did not include loading
from OQ3.

This corrects the table to label those rows as "partially supported", to
clarify the note (which had caused confusion), and to explain that
loading from OQ3 is not typically necessary for full-path submission in
the majority case.