Reorder cells in "Real-time benchmarking for qubit selection" tutorial (#3747)
The code cell corresponding to the sentence
>To benchmark the difference in performance, we consider a circuit that
prepares a Bell state across a linear chain of varying length. The
fidelity of the Bell state at the ends of the chain is measured.

is placed in a later location, seemingly out of place. This PR moves the
code cell to be immediately after the sentence.