Document macOS x86_64 downgraded to tier 2
In Qiskit 2.3 the macOS x86_64 platform will be downgraded from tier 1
to tier 2 (see Qiskit/Qiskit#15041 for more details). This is due to
lack of support from apple and limited resources available on Github
for the platform. This commit updates the documentation page for
platform support to reflect this change. This should not be merged
until Qiskit 2.3.0 as this change only takes effect in 2.3.0 and
2.2.x still has tier 1 support for the platform.