Fix Qiskit broken cross-references (#3968)
This PR fixes four broken cross-references in the Qiskit 2.2
`circuit_classical` page, and in the 1.1 & 2.1 release notes. Given that
this change impacts some unsupported versions of Qiskit, we can't open a
PR in the Qiskit repo to apply all the changes (see
https://github.com/Qiskit/qiskit/pull/15078 and
https://github.com/Qiskit/qiskit/pull/15080). Instead, I manually fixed
the HTML files in the Box artifacts and re-gerenerated the docs.

The only changes that can be synced into Qiskit/qiskit is from the
`circuit_classical` page
([PR](https://github.com/Qiskit/qiskit/pull/15079))

This change is important to avoid any potential issue with translations