Fixed `remove_instruction_from_target` in the Fractional Gate tutorial (#4257)
The error we would get before:
`qiskit.transpiler.exceptions.TranspilerError: "An instruction added
globally by class can't have properties set.`

This is due to the devices starting to emit the if_else instruction. As
a result, this tutorial currently fails due to this instruction. This
pull request introduces a fix to ensure compatibility by simply adding
the check of:

```
if qarg_map == {None: None}:
    qarg_map = None
```

Other fixes

- Also generalizes the info when selecting instance and backend
- Added requirement of Qiskit Basis Constructor (`pip install
qiskit_basis_constructor`)
- Switched to `least_busy` for selecting backends and made a note that
the backends used for this tutorial should be heron devices or later