Test contributor tooling on Windows (#4274)
Prework for #3233

This PR splits out the checks for our Python-based contributor tooling
so we can test them on Windows, Linux, and MacOS using a GitHub Actions
matrix job. Windows testing is important, even if we choose not to fully
support it because we want to make sure we're displaying an appropriate
error message.

Removing these tests from the main CI checks should also make CI
slightly faster (~30s) for other contributions.