Add more information to external link checker logs (#4282)
It's helpful to be able to glance at the logs and determine why a link
failed.

For example: A 502 status code is most likely a temporary flake, which
we can ignore. Whereas a 410 status code means the link is definitely
broken.

This PR adds the status code to the logs and reports if the fetch failed
for some other reason.