Fix some C API `objects.inv` links (#4304)
Part of https://github.com/Qiskit/documentation/issues/4200

Looking at the broken links, the vast majority have a period (`.`) in
their anchors because they're links to members of an object.

* One example is
`/docs/api/qiskit-c/qk-bit-term#qk_bitterm_label.bit_term`. The part
after the period is a parameter of the function
[`qk_bitterm_label`](https://quantum.cloud.ibm.com/docs/en/api/qiskit-c/qk-bit-term#qk_bitterm_label).
Parameters don't have HTML anchors, hence the broken link.
* Another example is
`/docs/api/qiskit-c/qk-circuit#qkdelayunit.qkdelayunit_ms`, a value of
[`QkDelayUnit`](https://quantum.cloud.ibm.com/docs/en/api/qiskit-c/qk-circuit#qkdelayunit).
This enum value _does_ have an anchor, but due to our conversion script,
the anchor is `#enumerator-qkdelayunit_ms`.

This PR fixes this by mapping such links to their parent objects. This
fixes 220/236 links. While it would be nice to be able to link directly
to the members, it's going to be quite a lot of work. Mapping to the
parent object is very simple (see
02758e8a4c224cb74d151dfc7d2a8b3344dc5c19), and will take the user very
close to the memeber.

I regenerated all APIs to make sure this didn't break any other links
(it fixed two links in `qiskit-c v2.1`). I will continue investigating
the other 16 broken links.

---------

Co-authored-by: Arnau Casau <47946624+arnaucasau@users.noreply.github.com>