Add single retry to external link checker (#4306)
Each week, we check over 2000 external links. Even if most links have a
small probability of flaking, we're likely to encounter them due to the
large number of links we check. For example, the most recent check
failed, but the reported links seemed to work fine for me:
https://github.com/Qiskit/documentation/issues/4301#issuecomment-3570114380.
When flakes do happen, an issue is created and a human has to manually
double-check it.

This PR adds a one-off retry to the link checker. We store the list of
broken links, which should be a small number (<~5), and retry them at
the end. If any link fails the retry, the checker reports a failure and
makes an issue like before. But if all links pass, no problems are
reported.

Hopefully this will reduce noise from the link checker.