Speed up internal link checker (#4326)
Improving check time is important to our team (see #2854). A large
portion of the time is spent in the internal link checker.

I profiled the link checker to see if there were any opportunities to
improve things. It looks like most of the time is spent parsing
markdown, which we can't really avoid. However, we can do a few things
to mitigate it without introducing any more complexity:


* c09865fd512d24ffc828d200ac985583136c7e1d does two things
* Remove the stringification step, we don't need to convert the AST back
to markdown, so we can avoid that work.
  * Traverse the tree one time, rather than three separate times
* 2b92c9ecac8727ea2209fde2354c23fab1e4b721 avoids blocking on file read.
This means we're parsing markdown while we're waiting for the
filesystem.


A very crude test shows somewhere between a 10%–30% speedup on my
machine.

Before: (30.53, 29.10, 29.08, 29.41) Mean 29.53 (+-~1) 
After:  (25.19, 23.84, 23.95, 25.56) Mean 24.64 (+-~1)

EDIT: More evidence this is a real speedup is the link checker in the
API checks takes ~6m for [this
PR](https://github.com/Qiskit/documentation/actions/runs/19760027099/job/56619546894),
whereas it took between 8–9m in the last three PRs that it ran on, which
roughly matches the speedup estimate on my machine.