Combine PR checks into single workflow (#4349)
This PR enables fixing #1970 by combining all PR checks into a single
workflow. We'll then need to edit the settings to require all jobs in
this workflow.

There are some things I'd like to neaten up in follow-up PRs, but this
is already a large PR to review, so I've just done the minimum to keep
the same behaviour as before while satisfying #1970.

---------

Co-authored-by: Eric Arellano <14852634+Eric-Arellano@users.noreply.github.com>