Fix CI for large PRs (#4461)
https://github.com/cli/cli/issues/10712 resulted in our Runtime PR
failing.

I debated using Git to get the information, but didn't because our
`fetch-depth` is 1 and we would need to fetch more, which isn't great
with our repository's size. I also don't want to use
`tj-actions/changed-files` due to supply chain security and because it
also under-the-hood Git fetches more commits.

So, we now use the API directly. The REST API has overly verbose
information with the actual patch for each file, so we use GraphQL.